Add files via upload
The Vehicle Parameter System IoT project is an Internet of Things (IoT) solution that aims to collect and monitor vehicle parameters in real time using various sensors and communicate the data to a centralized server or cloud platform. This project involves integrating multiple hardware components and software systems to achieve this objective.
The hardware components used in this project include various sensors such as GPS, temperature sensors, pressure sensors, and accelerometer sensors, which are used to measure different vehicle parameters such as speed, temperature, pressure, and acceleration. These sensors are connected to a microcontroller unit, such as an Arduino or Raspberry Pi, which is responsible for collecting the sensor data and processing it for further transmission.
The software systems used in this project include the firmware running on the microcontroller, which is responsible for collecting and processing the sensor data. The firmware communicates with a wireless communication module, such as GSM or Wi-Fi, to transmit the data to a centralized server or cloud platform for further analysis.
The centralized server or cloud platform receives the data from multiple vehicles and processes it for various applications, such as vehicle performance analysis, predictive maintenance, and route optimization. The data can also be used to alert the vehicle owner or driver in case of any abnormal conditions, such as overheating or low fuel levels.
Overall, the Vehicle Parameter System IoT project aims to provide an efficient and cost-effective solution for monitoring vehicle parameters in real time and using this data for various applications.